Delete Ingress file from template and ingress configs from values, also add listeningNodePort to values
The reason is that first of all the protocol is neigher HTTP or HTTPS so the most common ingress which is Nginx-Ingress can not handle that. It also misleads others so they think the chart may contain some sort of dashboard on port 80 which it does not. There was port:80 in values which never used since the container never opens port 80.

I also added listeningNodePort value so if the NodePort type is selected for the service type, the user can easily set the NodePort otherwise for ClusterIP or LoadBalancer types it will be skipped